requirements.txt
amqp==5.1.0
apscheduler == 3.9.1
asgiref==3.5.0
billiard==3.6.4.0
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.0.4
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.4
Django==4.0.4
django-apscheduler==0.6.2
django-crispy-forms==1.14.0
djangorestframework==3.13.1
gunicorn==20.1.0
idna==3.3
kombu==5.2.4
mpmath==1.2.1
prompt-toolkit==3.0.28
psycopg2==2.9.3
PyQRCode==1.2.1
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
requests==2.27.1
six==1.16.0
sqlparse==0.4.2
sympy==1.10
tzdata==2021.5
tzlocal==4.1
urllib3==1.26.9
vine==5.0.0
wcwidth==0.2.5
whitenoise==6.0.0